+++ MicroServices/loan-application/src/main/java/com/olbs/loan/service/iLoanService.java
@@ -9,8 +9,8 @@ import com.olbs.loan.entity.Loan;
 
 public interface iLoanService {
 	public Loan applyLoan(LoanRequestDto l);
-	public List<Loan> getLoansByCustId(int custId);
-	public Loan closeLoan(int loanId);
+	public List<Loan> getLoansByCustId(Long custId);
+	public Loan closeLoan(Long loanId);
 	public Customer_Info addCustomner(Customer_Info c);
 	public Account_info emiPayment(LoanRequestDto loan);
 }
